added rest of endpoints

+++ sensorsGenerator/src/main/java/es/p50/sensorsGenerator/co2/CO2.java
@@ -8,9 +8,9 @@ public class CO2 {
     private String unit;
     private String unitAbreviation;
     private double value;
-    private int timestamp;
+    private long timestamp;
 
-    public CO2(int id, String dataType, String sensorType, String unit, String unitAbreviation, int value, int timestamp) {
+    public CO2(int id, String dataType, String sensorType, String unit, String unitAbreviation, int value, long timestamp) {
         this.id = id;
         this.dataType = dataType;
         this.sensorType = sensorType;
@@ -68,11 +68,11 @@ public class CO2 {
         this.value = (double) value/10.0;
     }
 
-    public int getTimestamp() {
+    public long getTimestamp() {
         return this.timestamp;
     }
 
-    public void setTimestamp(int timestamp) {
+    public void setTimestamp(long timestamp) {
         this.timestamp = timestamp;
     }
